--- a/compiler/src/iree/compiler/Codegen/Common/GPU/GPUDistributionPatterns.cpp
+++ b/compiler/src/iree/compiler/Codegen/Common/GPU/GPUDistributionPatterns.cpp
@@ -161,6 +161,14 @@ struct DistributeScfFor final : OpDistributionPattern<scf::ForOp> {
     newForOp->setAttrs(forOp->getAttrs());
     Block *loopBody = newForOp.getBody();
 
+    // If there are no iterator arguments, the scf.for builder creates a
+    // terminator... This is really inconsistent and should be ideally fixed
+    // upstream. Erase the terminator, since we will merge a new terminator from
+    // the old block anyway.
+    if (forOp.getInitArgs().empty()) {
+      rewriter.eraseOp(loopBody->getTerminator());
+    }
+
     // Set up new iter_args. The loop body uses SIMD, so wrap the SIMD iter_args
     // of the new loop op into ToSIMDOps.
     rewriter.setInsertionPointToStart(loopBody);