@Groverkss 2025-03-10
> Just test llama3 benchmark on input size length 128/2048 by locally rebased this PR on 3.3.0rc20250310. Improve performance around **1ms**. 128: 37.2 ms->36.4 ms. 2048: 174 ms -> 173 ms.

This should have no affect on any benchmarks...