@Groverkss 2025-05-02
> @Groverkss Is this still under development or is it stale? It can always be reopened if development resumes.

It's not stale, I plan to get back on this next week. It needs some more changes in IREE before I can make this not a draft. I'd like to keep this open since it's a good tracking pr for me.